* changed Flash Version to 9 (was 10 temporarily but MP3Pitch didn't work out) and framerate to 60 * added placeholder MenuState to be renamed and developed into Player selection screen next time * sanitized debug messsages from Spring
!! TODO: gameplay goddamnit !!

IDEAS:
concept for final state layout:
exactly like original with addition of optionselectstate:

OptionSelectState => PlayerSelectState => PlayState => ScoreState

PlayerSelectState
Load directly into the PlayerSelectState with a little text at the bottom, "[ESC] select level / change game mode" or something to that effect (ESC in black with a white rounded outline) amount of players jump over log and touch the right edge to begin PlayState.

OptionSelectState
if ESC is pressed, go to optionSelectState where from a grid of thumbnails a level can be chosen and below the game mode (deathmatch, lord of the flies (after the first kill, killer is the Lord of the Flies (like the cheat), if he is killed, the new killer becomes the lord. score counts up for each second that a player is the lord, win at 99( setting? ), beeping for the last five seconds?)
* options are mouse based?